Fixes #26241 - revert to pulp api for paths

+++ app/services/katello/pulp3/repository.rb
@@ -113,11 +113,10 @@ module Katello
 
       def delete_distributions
         paths.values.each do |path|
+          dists = lookup_distributions(base_path: path.sub(/^\//, ''))
+          delete_distribution(dists.first._href) if dists.first
           distribution_reference = distribution_reference(path)
-          if distribution_reference
-            delete_distribution(distribution_reference.href)
-            distribution_reference.destroy
-          end
+          distribution_reference.destroy if distribution_reference
         end
       end
